change: 自动排期算法提取
