+++ app/components/Admin/ScheduleManager.vue
@@ -1590,6 +1590,7 @@ import { convertToHttps, getNeteaseCookie } from '~/utils/url'
 import { useLocale } from '~/utils/locale'
 import { useServerErrors } from '~/composables/useLocaleText'
 import { formatDuration, addDaysToString, getDaysBetween } from '~/utils/timeUtils'
+import { autoSchedule, poolCandidateFromItem } from '~/utils/autoSchedule'
 
 import SchedulePlaylistFilterModal from './SchedulePlaylistFilterModal.vue'
 import { getPlaylistDetail } from '~/utils/neteaseApi'
@@ -3171,115 +3172,6 @@ const resetAutoSchedule = () => {
   autoScheduleResult.value = { songs: [], totalDuration: 0, diff: 0, absDiff: 0 }
 }
 
-/**
- * 自动排期算法：在候选池中选取总时长最接近目标的歌曲组合
- * 策略：主路径按降序贪心，辅路径随机采样，取更优结果
- * @param {'under'|'over'} direction 'under' 总时长不超过目标，'over' 总时长不低于目标
- */
-const autoSchedule = (direction, candidates) => {
-  const targetSeconds = Math.floor(autoScheduleTargetMinutes.value * 60)
-  const filteredCandidates = candidates
-    .filter((s) => typeof s.durationSeconds === 'number' && s.durationSeconds > 0)
-
-  if (filteredCandidates.length === 0) {
-    return { songs: [], totalDuration: 0, diff: 0, absDiff: 0 }
-  }
-
-  // 建立 id → song 的 Map，避免随机路径中 O(n²) 查找
-  const songMap = new Map(filteredCandidates.map((s) => [s.id, s]))
-
-  const runGreedy = (sorted) => {
-    const result = []
-    let total = 0
-    for (const song of sorted) {
-      const newTotal = total + song.durationSeconds
-      if (direction === 'under') {
-        if (newTotal <= targetSeconds) {
-          result.push(song)
-          total = newTotal
-        }
-      } else {
-        result.push(song)
-        total = newTotal
-        if (total >= targetSeconds) break
-      }
-    }
-    if (direction === 'under') {
-      const ids = new Set(result.map((s) => s.id))
-      for (const song of filteredCandidates.slice().sort((a, b) => a.durationSeconds - b.durationSeconds)) {
-        if (ids.has(song.id)) continue
-        if (total + song.durationSeconds <= targetSeconds) {
-          result.push(song)
-          total += song.durationSeconds
-          ids.add(song.id)
-        }
-      }
-    } else {
-      // 'over' 方向回填：尝试用更短的歌曲替换已选歌曲，在保持总时长 >= 目标的前提下尽量逼近
-      const selectedIds = new Set(result.map((s) => s.id))
-      const remaining = filteredCandidates.filter((s) => !selectedIds.has(s.id))
-        .sort((a, b) => a.durationSeconds - b.durationSeconds)
-      // 从最短的已选歌曲开始尝试替换（越短越有可能被替换掉仍保持 >= target）
-      for (let ri = result.length - 1; ri >= 0; ri--) {
-        const currentSong = result[ri]
-        const currentDuration = currentSong.durationSeconds
-        const totalWithoutThis = total - currentDuration
-        const minReplacement = targetSeconds - totalWithoutThis
-        if (minReplacement >= currentDuration) continue
-        for (let ii = 0; ii < remaining.length; ii++) {
-          const candidate = remaining[ii]
-          if (candidate.durationSeconds >= minReplacement && candidate.durationSeconds < currentDuration) {
-            result[ri] = candidate
-            total = totalWithoutThis + candidate.durationSeconds
-            selectedIds.delete(currentSong.id)
-            selectedIds.add(candidate.id)
-            remaining.splice(ii, 1)
-            break
-          }
-        }
-      }
-    }
-    const diff = total - targetSeconds
-    return { songs: result, totalDuration: total, diff, absDiff: Math.abs(diff) }
-  }
-
-  const r1 = runGreedy([...filteredCandidates].sort((a, b) => b.durationSeconds - a.durationSeconds))
-
-  const shuffled = filteredCandidates.map((s) => s.id)
-  for (let i = shuffled.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1))
-    ;[shuffled[i], shuffled[j]] = [shuffled[j], shuffled[i]]
-  }
-  const r2 = runGreedy(shuffled.map((id) => songMap.get(id)).filter(Boolean))
-
-  // tiebreaker：absDiff 相同时优先取歌曲数较多的结果
-  const pick = r1.absDiff < r2.absDiff
-    ? r1
-    : r1.absDiff > r2.absDiff
-      ? r2
-      : r1.songs.length >= r2.songs.length
-        ? r1
-        : r2
-  return pick
-}
-
-// 自动排期候选池：根据选择的来源返回待排库/备选池/合并的歌曲
-// 备选池条目统一构造器（含 replayRequestId，用于重播申请链路）
-function poolCandidateFromItem(item) {
-  return {
-    id: item.songId,
-    songId: item.songId,
-    title: item.title,
-    artist: item.artist,
-    durationSeconds: item.durationSeconds,
-    replayRequestId: item.replayRequestId || null,
-    musicId: item.musicId,
-    musicPlatform: item.musicPlatform,
-    requester: item.requester,
-    cover: item.cover,
-    createdAt: item.createdAt
-  }
-}
 const autoScheduleCandidates = computed(() => {
   const scheduledIds = new Set(localScheduledSongs.value.map((s) => s.song && s.song.id).filter(Boolean))
 
@@ -3305,9 +3197,8 @@ const autoScheduleCandidates = computed(() => {
 
 const runAutoSchedule = () => {
   const candidates = autoScheduleCandidates.value
-  // 快照候选集 ID，供 confirmAutoSchedule 过滤使用，避免用户切换 Tab / 更新搜索导致候选集变化而静默丢弃结果
-  const candidateIds = new Set(candidates.map((s) => s.id))
-  const result = autoSchedule(autoScheduleDirection.value, candidates)
+  const candidateIds = new Set(candidates.map((s) => s.id)) // 快照候选集 ID，确认时用于过滤
+  const result = autoSchedule(autoScheduleDirection.value, autoScheduleTargetMinutes.value, candidates)
   if (result.songs.length === 0) {
     if (window.$showNotification) {
       window.$showNotification(callLocale('autoScheduleNoResult', '未能找到满足条件的歌曲组合'), 'warning')
